Split BattleService: extract offline replay and zone resolution (#1516)
BattleService had grown to ~940 lines with 14 constructor deps hosting
three responsibilities: live battle lifecycle, offline/switch replay
orchestration, and zone viability/relocation. Extract the latter two
into their own services, mirroring the existing ChallengeRewardService/
ProficiencyRewardService pattern:

- ZoneResolutionService: viability, lazy relocation, unlock checks, and
  challengeable-boss resolution — shared by BattleService's live paths
  and OfflineProgressService's away-window replay.
- OfflineProgressService: SimulateOfflineProgress/SimulateSwitchProgress
  and the reward-consolidation pass, touching the live lifecycle only
  through BattleService.ResolveStaleBattle.

Pure refactor, no behavior change — existing integration tests (renamed
to OfflineProgressServiceIntegrationTests) pin the seams.

Closes #1516
