+++ Game.Application/Services/BattleService.cs
@@ -4,14 +4,12 @@ using Game.Abstractions.DataAccess;
 using Game.Core;
 using Game.Core.Attributes;
 using Game.Core.Battle;
-using Game.Core.Battle.Offline;
 using Game.Core.Players;
 using Game.Core.Proficiencies;
 using Game.Core.Progress;
 using Microsoft.Extensions.Logging;
 using CoreClass = Game.Core.Classes.Class;
 using CoreEnemy = Game.Core.Enemies.Enemy;
-using CoreZone = Game.Core.Zones.Zone;
 
 namespace Game.Application.Services
 {
@@ -26,9 +24,7 @@ namespace Game.Application.Services
         IProficiencies proficiencies,
         IClasses classes,
         BattleFactory battleFactory,
-        OfflineProgressSimulator offlineSimulator,
-        ChallengeRewardService challengeRewards,
-        ProficiencyRewardService proficiencyRewards,
+        ZoneResolutionService zoneResolution,
         ILogger<BattleService> logger)
     {
         private readonly IPlayerRepository _playerRepo = playerRepo;
@@ -41,22 +37,8 @@ namespace Game.Application.Services
         private readonly IProficiencies _proficiencies = proficiencies;
         private readonly IClasses _classes = classes;
         private readonly BattleFactory _battleFactory = battleFactory;
-        private readonly OfflineProgressSimulator _offlineSimulator = offlineSimulator;
-        private readonly ChallengeRewardService _challengeRewards = challengeRewards;
-        private readonly ProficiencyRewardService _proficiencyRewards = proficiencyRewards;
+        private readonly ZoneResolutionService _zoneResolution = zoneResolution;
         private readonly ILogger<BattleService> _logger = logger;
-
-        // Offline-rewards window bounds (spike #879). Below the minimum, a return is treated as no time away
-        // (no rewards, just re-anchor); above the cap, only the cap is ever simulated so a long absence (or an
-        // all-draw zone) can't run unbounded CPU.
-        public static readonly TimeSpan MinimumOfflineAway = TimeSpan.FromMinutes(5);
-        public static readonly TimeSpan MaximumOfflineSimulation = TimeSpan.FromHours(10);
-
-        // CPU-waste guard handed to the offline simulator: an opening batch of pure draws is a stalemate the
-        // player can neither win nor lose, so simulating the whole away budget of maximum-duration draws earns
-        // nothing for the most work. Stop after this many opening all-draw battles (a true stalemate draws
-        // every time, so a small count is decisive; any win or loss disables the guard).
-        private const int StalemateCutoffBattles = 10;
 
         // Slack on the server-measured elapsed-time victory check: one logical tick, because the frontend's
         // battle-start may sit up to a tick off the backend's (a battle started mid-tick counts its first
@@ -67,8 +49,10 @@ namespace Game.Application.Services
 
         // Post-battle enemy cooldown, shared by the win and loss paths so the two cannot diverge. Both anchor
         // it to the server clock; the win path anchors to the battle's completion (battle start + replayed
-        // duration) and the loss path to the moment of the loss, but the duration is identical.
-        private static readonly TimeSpan PostBattleCooldown = TimeSpan.FromSeconds(5);
+        // duration) and the loss path to the moment of the loss, but the duration is identical. Internal (not
+        // private): OfflineProgressService's simulated loop uses the identical cooldown between battles so the
+        // offline pass paces the same as the live idle loop.
+        internal static readonly TimeSpan PostBattleCooldown = TimeSpan.FromSeconds(5);
 
         public async Task<BattleStartResult> StartBattle(Player player, PlayerState state, int zoneId, int? newZoneId = null, DateTime? scheduledStartTime = null, int? clientBattleMs = null, CancellationToken cancellationToken = default)
         {
@@ -89,7 +73,7 @@ namespace Game.Application.Services
                 var targetZone = _zones.GetDomainZone(newZoneId.Value);
                 if (!_zones.IsZoneRetired(newZoneId.Value)
                     && !_zones.IsHomeZone(newZoneId.Value)
-                    && await IsZoneUnlocked(player.Id, targetZone, cancellationToken))
+                    && await _zoneResolution.IsZoneUnlocked(player.Id, targetZone, cancellationToken))
                 {
                     player.ChangeZone(newZoneId.Value);
                     zoneId = newZoneId.Value;
@@ -100,7 +84,7 @@ namespace Game.Application.Services
             // Lazy relocation: if the resolved zone is no longer viable (it was retired, or every enemy
             // assigned to it has been retired), move the player to the nearest viable zone so the idle loop
             // never stalls on a non-navigable zone or throws spawning from an empty table.
-            zoneId = await EnsureViableZone(player, zoneId, cancellationToken);
+            zoneId = await _zoneResolution.EnsureViableZone(player, zoneId, cancellationToken);
 
             var zone = _zones.GetDomainZone(zoneId);
 
@@ -183,7 +167,7 @@ namespace Game.Application.Services
             // bossless, retired, or locked zone must be a true no-op rather than silently ending the player's
             // in-progress fight. A legitimate client can never be in a locked or retired zone to begin with
             // (the idle loop relocates out of a retired one), so those gates only block tampered requests.
-            var zone = await ResolveChallengeableBossZone(player.Id, zoneId, cancellationToken);
+            var zone = await _zoneResolution.ResolveChallengeableBossZone(player.Id, zoneId, cancellationToken);
             if (zone is null)
             {
                 return null;
@@ -197,7 +181,7 @@ namespace Game.Application.Services
             var now = DateTime.UtcNow;
             var seed = CreateBattleSeed();
 
-            var enemy = _battleFactory.CreateBossEnemy(zone, BossEnemyResolver(zone));
+            var enemy = _battleFactory.CreateBossEnemy(zone, _zoneResolution.BossEnemyResolver(zone));
 
             var enemySkillIds = enemy.BattleSkills.Select(skill => skill.Id).ToList();
             var snapshot = BattleSnapshot.FromPlayer(player, await CaptureProficiencyLevels(player.Id, cancellationToken));
@@ -232,7 +216,7 @@ namespace Game.Application.Services
             // Anti-cheat: a retired, locked, or bossless current zone cannot be boss-farmed. The zone is the
             // player's own CurrentZoneId (not client-supplied), so a tampered client can't farm a zone it
             // isn't in; the meaningful gate is that the current zone actually has a challengeable boss.
-            if (await ResolveChallengeableBossZone(player.Id, player.CurrentZoneId, cancellationToken) is null)
+            if (await _zoneResolution.ResolveChallengeableBossZone(player.Id, player.CurrentZoneId, cancellationToken) is null)
             {
                 return false;
             }
@@ -358,229 +342,6 @@ namespace Game.Application.Services
             return AbandonBattle(player, state, cancellationToken: cancellationToken);
         }
 
-        /// <summary>
-        /// Computes how long the player was away, replays the missed idle/boss battles (spike #879), applies
-        /// the accumulated rewards, and returns a welcome-back summary. Runs inline within the calling socket
-        /// command (the cap keeps the worst case well under the command timeout); the command's cancellation
-        /// token is threaded into the simulation loop so a long run unwinds promptly.
-        /// <para>
-        /// Away time is <c>now − <see cref="Player.LastActivity"/></c>, measured server-side. Below
-        /// <see cref="MinimumOfflineAway"/> it is a no-op beyond re-anchoring <c>LastActivity</c> (so a second
-        /// immediate call earns nothing); otherwise it resolves any stale in-flight battle, resumes the
-        /// persisted loop mode in the player's (viability-checked) current zone, simulates up to
-        /// <see cref="MaximumOfflineSimulation"/>, applies exp per victory and the consolidated
-        /// statistics/challenges, re-anchors <c>LastActivity</c>, and persists once.
-        /// </para>
-        /// </summary>
-        public Task<OfflineProgressSummary> SimulateOfflineProgress(Player player, PlayerState state, CancellationToken cancellationToken = default)
-        {
-            return SimulateProgress(player, state, MinimumOfflineAway, cancellationToken);
-        }
-
-        /// <summary>
-        /// Credits a character being switched away from in a deliberate in-game character switch (spike #922):
-        /// the same elapsed-time replay as <see cref="SimulateOfflineProgress"/> but with the
-        /// <see cref="MinimumOfflineAway"/> floor dropped, so any elapsed time since the departed character's
-        /// <see cref="Player.LastActivity"/> is credited (the 5-minute floor is a login-time concern). Resolves
-        /// the in-flight battle, applies the rewards, re-anchors <c>LastActivity</c>, and persists — so the
-        /// departed character loses no idle progress when the player switches to another of their characters.
-        /// The caller must invoke this off the departed character's battle loop (its socket torn down first),
-        /// so the player-state write cannot race a live battle-completion command.
-        /// </summary>
-        public Task<OfflineProgressSummary> SimulateSwitchProgress(Player player, PlayerState state, CancellationToken cancellationToken = default)
-        {
-            return SimulateProgress(player, state, TimeSpan.Zero, cancellationToken);
-        }
-
-        // Shared elapsed-time replay for both the login welcome-back path and the deliberate-switch credit. The
-        // only difference is the away floor: the login path skips a sub-5-minute return, while a switch credits
-        // any elapsed time (minimumAway == zero), so the two cannot otherwise drift.
-        private async Task<OfflineProgressSummary> SimulateProgress(Player player, PlayerState state, TimeSpan minimumAway, CancellationToken cancellationToken)
-        {
-            var now = DateTime.UtcNow;
-            var awayMs = (long)(now - player.LastActivity).TotalMilliseconds;
-            var cappedAwayMs = Math.Min(awayMs, (long)MaximumOfflineSimulation.TotalMilliseconds);
-
-            // Below the threshold there are no offline rewards. Re-anchor LastActivity (so the next away period
-            // starts fresh and an immediate re-claim is a no-op) and return an empty summary. Any stale
-            // in-flight battle is left for the idle loop's first StartBattle to abandon, exactly as on a normal
-            // reconnect — there is no away window to simulate, so settling it here would change nothing.
-            if (awayMs < minimumAway.TotalMilliseconds)
-            {
-                player.StampActivity(now);
-                await _playerRepo.SavePlayer(player, cancellationToken);
-                return OfflineProgressSummary.Empty(cappedAwayMs, player.AutoChallengeBoss, player.CurrentZoneId);
-            }
-
-            // Settle the disconnected battle before simulating the away window, so its outcome is credited once
-            // (here) rather than being re-abandoned by the first live StartBattle after the gate.
-            await ResolveStaleBattle(player, state, cancellationToken);
-
-            // Load the progress aggregate once for the whole offline pass. Its completed challenges (loop/zone
-            // gating), proficiency levels (battle snapshot), and statistics (reward application) all read the
-            // same Progress_{playerId} cache key, so threading one aggregate down replaces 3-4 serial round-trips.
-            // Loaded after ResolveStaleBattle so it reflects the settled stale battle's progress mutation.
-            var progress = await _progressRepo.Load(player, cancellationToken);
-            var completedChallengeIds = progress.CompletedChallengeIds();
-
-            var (mode, zone) = await ResolveOfflineLoop(player, completedChallengeIds, cancellationToken);
-            var proficiencyLevels = ToProficiencyLevels(progress.Proficiencies);
-            var parameters = BuildSimulationParameters(player, mode, zone, awayMs, proficiencyLevels);
-
-            var result = _offlineSimulator.Simulate(parameters, cancellationToken);
-
-            var levelBefore = player.Level;
-            var statPointsBefore = player.StatPoints.StatPointsGained;
-            var rewards = await ApplyOfflineRewards(player, progress, result, cancellationToken);
-
-            // Re-anchor the away clock and persist the player (exp/levels/unlocks) in one save. The exp batch
-            // already raised its own single core update in ApplyOfflineRewards; this re-anchor raises one more.
-            // Both are absolute write-behind writes (the final state persists regardless), and two is still
-            // nowhere near the per-victory flood decision 6 avoids.
-            player.StampActivity(now);
-            await _playerRepo.SavePlayer(player, cancellationToken);
-
-            return new OfflineProgressSummary
-            {
-                AwayMs = cappedAwayMs,
-                AutoChallengeBoss = result.IsBossBattle,
-                ZoneId = result.ZoneId,
-                BattlesWon = result.Wins,
-                BattlesLost = result.Losses,
-                BattlesDrawn = result.Draws,
-                TotalExp = result.TotalExp,
-                LevelsGained = player.Level - levelBefore,
-                StatPointsGained = player.StatPoints.StatPointsGained - statPointsBefore,
-                CompletedChallenges = rewards.CompletedChallenges,
-                ProficiencyGains = rewards.ProficiencyGains.Results,
-                OpenedProficiencies = rewards.ProficiencyGains.Opened,
-            };
-        }
-
-        // Resolves which loop the offline simulation resumes and the zone it runs in. Boss mode resumes only
-        // when the persisted flag is set and the current zone still has a challengeable boss (in circulation,
-        // unlocked, boss authored); otherwise the loop falls back to idle in the nearest viable zone (the same
-        // lazy-relocation the live idle loop uses, so a retired/empty current zone never stalls the sim).
-        private async Task<(OfflineLoopMode Mode, CoreZone Zone)> ResolveOfflineLoop(
-            Player player, IReadOnlySet<int> completedChallengeIds, CancellationToken cancellationToken)
-        {
-            var currentZoneId = player.CurrentZoneId;
-            if (player.AutoChallengeBoss
-                && ResolveChallengeableBossZone(currentZoneId, completedChallengeIds) is { } bossZone)
-            {
-                return (OfflineLoopMode.Boss, bossZone);
-            }
-
-            var idleZoneId = await EnsureViableZone(player, currentZoneId, completedChallengeIds, cancellationToken);
-            return (OfflineLoopMode.Idle, _zones.GetDomainZone(idleZoneId));
-        }
-
-        // Builds the simulator inputs for the resolved loop. Idle rolls a random per-zone spawn each battle;
-        // boss builds the zone's dedicated boss deterministically — the same factory resolvers the live battle
-        // start uses, so an offline battle is constructed identically to an online one. The player snapshot and
-        // the catalog resolvers are shared across the whole run (the player's power is stationary offline).
-        private OfflineSimulationParameters BuildSimulationParameters(
-            Player player, OfflineLoopMode mode, CoreZone zone, long awayMs,
-            IReadOnlyList<ProficiencyLevelSnapshot> proficiencyLevels)
-        {
-            Func<int, CoreEnemy> resolveEnemy = mode == OfflineLoopMode.Boss
-                ? BossEnemyResolver(zone)
-                : level => _enemies.GetRandomDomainEnemy(zone.Id, level);
-
-            return new OfflineSimulationParameters
-            {
-                // One snapshot drives the whole window: the player's power — proficiency levels included — is
-                // frozen at the window start, so the away period fights at a stationary power even as the
-                // simulated victories accrue proficiency XP (mirroring how gear and stats are frozen).
-                Snapshot = BattleSnapshot.FromPlayer(player, proficiencyLevels),
-                Mode = mode,
-                Zone = zone,
-                AwayBudgetMs = awayMs,
-                CapMs = (long)MaximumOfflineSimulation.TotalMilliseconds,
-                CooldownMs = (int)PostBattleCooldown.TotalMilliseconds,
-                ResolveEnemy = resolveEnemy,
-                ResolveItem = _items.GetItem,
-                ResolveMod = _itemMods.GetItemMod,
-                ResolveSkill = _skills.TryGetSkill,
-                ResolveProficiency = _proficiencies.GetProficiency,
-                ResolveClass = ResolveClass,
-                SeedSource = CreateBattleSeed,
-                StalemateCutoffBattles = StalemateCutoffBattles,
-            };
-        }
-
-        // The dedicated-boss resolver shared by the live boss challenge and the offline boss loop: the zone's
-        // authored boss at the requested (fixed boss) level. Callers only reach this for a zone the
-        // challengeable-boss gate resolved (BossEnemyId set), so the null check here is a defensive invariant
-        // rather than a reachable state.
-        private Func<int, CoreEnemy> BossEnemyResolver(CoreZone zone)
-        {
-            var bossEnemyId = zone.BossEnemyId
-                ?? throw new InvalidOperationException($"Zone {zone.Id} has no dedicated boss enemy authored.");
-            return level => _enemies.GetDomainEnemy(bossEnemyId, level)
-                ?? throw new InvalidOperationException(
-                    $"Zone {zone.Id} references boss enemy {bossEnemyId}, which does not exist.");
-        }
-
-        // Applies a simulated away window's rewards to the player and progress in one consolidated pass
-        // (spike #879 decisions 6 & 7): each battle feeds the same per-battle statistics path the live handler
-        // uses, exp is granted per victory (so the per-grant clamp never truncates a haul and levels accrue),
-        // and the affected challenges are evaluated once at the end with the live per-challenge push suppressed
-        // (the summary is the notification). Returns the completed challenges and the folded proficiency gains
-        // (spike #982 decision 9 — the offline accrual's notification rides the summary, not a per-battle push).
-        private async Task<OfflineRewards> ApplyOfflineRewards(
-            Player player, PlayerProgress progress, OfflineProgressResult result, CancellationToken cancellationToken)
-        {
-            if (result.BattlesSimulated == 0)
-            {
-                return OfflineRewards.Empty;
-            }
-
-            var victoryExpRewards = new List<int>();
-            var proficiencyGains = new ProficiencyGainAccumulator();
-            // Union the statistic rows touched across every battle, so the end-of-window challenge evaluation
-            // sees every moved statistic. RecordBattleCompleted currently returns the progress aggregate's
-            // cumulative dirty set (it is loaded once for the whole window), but unioning each call's result
-            // explicitly keeps this correct even if that method were ever changed to return only the per-battle
-            // delta — a mixed-outcome window can touch a tracked statistic in a non-final battle (e.g. a kill
-            // challenge crossed by early wins before a closing loss), and the union captures it regardless.
-            var touchedStatistics = new HashSet<(EStatisticType Type, int? EntityId)>();
-            foreach (var battle in result.Battles)
-            {
-                foreach (var key in progress.RecordBattleCompleted(
-                    battle.Enemy, battle.Result.Victory, battle.Result.PlayerDied, battle.Result.TotalMs,
-                    battle.Result.Stats, result.IsBossBattle, result.ZoneId))
-                {
-                    touchedStatistics.Add(key);
-                }
-
-                if (battle.Result.Victory)
-                {
-                    victoryExpRewards.Add(battle.ExpReward);
-
-                    // Accrue proficiency XP per won battle, exactly as the live handler does — same inputs
-                    // (this battle's skill stats + player power), same service — so the offline accrual matches
-                    // what the player would have earned live (the "offline == live" invariant). The push is
-                    // suppressed; the folded results ride the welcome-back summary instead.
-                    proficiencyGains.Add(_proficiencyRewards.AccrueAndApply(
-                        progress, battle.Result.Stats, battle.PlayerPower, player, notify: false));
-                }
-            }
-
-            // Grant the whole window's exp before evaluating challenges, so a statistic-independent
-            // LevelReached challenge sees the post-window level (mirroring the live order, where exp is granted
-            // before the battle-completed handler evaluates challenges).
-            if (victoryExpRewards.Count > 0)
-            {
-                player.GrantOfflineExp(victoryExpRewards);
-            }
-
-            var completed = _challengeRewards.EvaluateAndApply(progress, touchedStatistics, player, notify: false);
-
-            await _progressRepo.Save(progress, cancellationToken);
-            return new OfflineRewards(completed, proficiencyGains.Build());
-        }
-
         private async Task AbandonBattle(Player player, PlayerState state, int? clientBattleMs = null, CancellationToken cancellationToken = default)
         {
             var now = DateTime.UtcNow;
@@ -677,113 +438,6 @@ namespace Game.Application.Services
             return rewards;
         }
 
-        // Whether a zone is viable for the idle loop: in circulation (not retired) and carrying at least one
-        // spawnable enemy. The two facts live in separate caches (zone retirement vs the enemy spawn tables),
-        // so they are combined here rather than on either lean domain model. The id is range-checked first so
-        // a stale/out-of-range CurrentZoneId reads as non-viable (and triggers relocation) instead of throwing.
-        private bool IsZoneViable(int zoneId)
-        {
-            return _zones.ValidateZoneId(zoneId)
-                && !_zones.IsZoneRetired(zoneId)
-                && _enemies.HasSpawnableEnemies(zoneId);
-        }
-
-        // Live-path overload: defers reading the completed-challenge set until a relocation is actually needed
-        // (the viability check short-circuits first), so the idle hot path never pays the read.
-        private async Task<int> EnsureViableZone(Player player, int zoneId, CancellationToken cancellationToken)
-        {
-            // An out-of-range zone id is corruption/tampering, not a relocation case: leave it for the
-            // downstream GetDomainZone to surface loudly (fail-fast) rather than silently relocating.
-            if (!_zones.ValidateZoneId(zoneId) || IsZoneViable(zoneId))
-            {
-                return zoneId;
-            }
-
-            var completedChallengeIds = await _progressRepo.GetCompletedChallengeIds(player.Id, cancellationToken);
-            return await EnsureViableZone(player, zoneId, completedChallengeIds, cancellationToken);
-        }
-
-        // Relocates the player when their resolved zone is no longer viable, returning the zone the battle
-        // should run in. "Nearest" is the lowest-Order zone the player has unlocked that is viable, falling
-        // back to the starting zone. A no-op (no save) when the current zone is already viable. Takes the
-        // completed-challenge set so a caller holding a loaded progress aggregate (the offline pass) gates
-        // without re-reading the progress cache key.
-        private async Task<int> EnsureViableZone(
-            Player player, int zoneId, IReadOnlySet<int> completedChallengeIds, CancellationToken cancellationToken)
-        {
-            if (!_zones.ValidateZoneId(zoneId) || IsZoneViable(zoneId))
-            {
-                return zoneId;
-            }
-
-            // Filter to viable candidates before ordering, then resolve each domain zone once (lazily, so the
-            // resolve stops at the first unlocked match) rather than re-resolving it inside the predicate.
-            var destination = _zones.All()
-                .Where(zone => IsZoneViable(zone.Id))
-                .OrderBy(zone => zone.Order)
-                .Select(zone => _zones.GetDomainZone(zone.Id))
-                .FirstOrDefault(zone => zone.IsUnlocked(completedChallengeIds));
-            var newZoneId = destination?.Id ?? NewPlayerFactory.StartingZoneId;
-
-            if (newZoneId != player.CurrentZoneId)
-            {
-                player.ChangeZone(newZoneId);
-                await _playerRepo.SavePlayer(player, cancellationToken);
-            }
-
-            return newZoneId;
-        }
-
-        // Shared challengeable-boss gate (StartBossBattle / SetAutoChallengeBoss / ResolveOfflineLoop): resolves
-        // the zone iff its boss can actually be challenged — the id is in range (checked before GetDomainZone,
-        // which throws on an out-of-range id), the zone is in circulation, and a dedicated boss is authored.
-        // The unlock check differs per caller, so it lives on the two overloads below.
-        private CoreZone? ResolveBossZone(int zoneId)
-        {
-            if (!_zones.ValidateZoneId(zoneId) || _zones.IsZoneRetired(zoneId))
-            {
-                return null;
-            }
-
-            var zone = _zones.GetDomainZone(zoneId);
-            return zone.BossEnemyId is null ? null : zone;
-        }
-
-        // Live-path gate: the unlock check reads the player's completed challenges through IsZoneUnlocked,
-        // incurred only once the cheaper range/retired/boss checks have passed.
-        private async Task<CoreZone?> ResolveChallengeableBossZone(int playerId, int zoneId, CancellationToken cancellationToken)
-        {
-            var zone = ResolveBossZone(zoneId);
-            if (zone is null || !await IsZoneUnlocked(playerId, zone, cancellationToken))
-            {
-                return null;
-            }
-
-            return zone;
-        }
-
-        // Offline-pass gate: unlocks against a completed-challenge set the caller already holds (the loaded
-        // progress aggregate), so the check re-reads no progress cache key.
-        private CoreZone? ResolveChallengeableBossZone(int zoneId, IReadOnlySet<int> completedChallengeIds)
-        {
-            var zone = ResolveBossZone(zoneId);
-            return zone is not null && zone.IsUnlocked(completedChallengeIds) ? zone : null;
-        }
-
-        // Whether a zone is unlocked for the player. An ungated zone is always open and pays no read cost;
-        // a gated zone costs one indexed completion lookup, incurred only on a real zone transition or a
-        // boss challenge (not per idle tick). The unlock rule itself lives on the domain Zone.
-        private async Task<bool> IsZoneUnlocked(int playerId, CoreZone zone, CancellationToken cancellationToken = default)
-        {
-            if (zone.UnlockChallengeId is null)
-            {
-                return true;
-            }
-
-            var completedChallengeIds = await _progressRepo.GetCompletedChallengeIds(playerId, cancellationToken);
-            return zone.IsUnlocked(completedChallengeIds);
-        }
-
         // Generates the simulation RNG seed from a cryptographic (non-time) entropy source. A wall-clock seed
         // (DateTime.Ticks) is monotonic and low-entropy in its low 32 bits — correlated and predictable between
         // battles — which makes it unsuitable as the shared starting point for the parity-identical crit/dodge
@@ -870,86 +524,4 @@ namespace Game.Application.Services
         public required uint Seed { get; set; }
     }
 
-    /// <summary>
-    /// The welcome-back summary of a returning player's offline progress: how long they were away (capped),
-    /// which loop ran, the battle tally, and the rewards earned (exp, levels, stat points, and the challenges
-    /// completed with what they unlocked). Returned by <see cref="BattleService.SimulateOfflineProgress"/> and
-    /// projected to the API model the client gate renders.
-    /// </summary>
-    public class OfflineProgressSummary
-    {
-        /// <summary>How long the player was away, in milliseconds, clamped to the simulation cap.</summary>
-        public required long AwayMs { get; init; }
-
-        /// <summary>Whether the simulated loop was auto-challenging the boss (<c>true</c>) or idle-farming
-        /// (<c>false</c>).</summary>
-        public required bool AutoChallengeBoss { get; init; }
-
-        /// <summary>The zone the loop ran in.</summary>
-        public required int ZoneId { get; init; }
-
-        public required int BattlesWon { get; init; }
-        public required int BattlesLost { get; init; }
-        public required int BattlesDrawn { get; init; }
-
-        /// <summary>Total experience earned across all victories.</summary>
-        public required long TotalExp { get; init; }
-
-        /// <summary>Levels gained over the window.</summary>
-        public required int LevelsGained { get; init; }
-
-        /// <summary>Stat points gained over the window (from the levels gained).</summary>
-        public required int StatPointsGained { get; init; }
-
-        /// <summary>The challenges completed over the window, each with the reward ids it unlocked.</summary>
-        public required IReadOnlyList<CompletedChallenge> CompletedChallenges { get; init; }
-
-        /// <summary>The proficiency gains accrued over the window, folded across every won battle: each trained
-        /// proficiency's total XP gained, its final level/residual XP, the milestones it crossed, and the reward
-        /// skills granted (spike #982 decision 9 — the offline accrual's notification rides this summary).</summary>
-        public required IReadOnlyList<ProficiencyXpResult> ProficiencyGains { get; init; }
-
-        /// <summary>The proficiency nodes opened over the window (a maxed tier's next tier or a newly-satisfied
-        /// gateway), each with the seed skill it granted (if any).</summary>
-        public required IReadOnlyList<ProficiencyOpened> OpenedProficiencies { get; init; }
-
-        /// <summary>
-        /// Whether the window produced anything worth gating on. The frontend skips the welcome-back gate for
-        /// an empty summary (a sub-threshold absence, or one that earned nothing) and enters the game directly.
-        /// A window that only advanced proficiencies (e.g. a maxed-XP-level character) still reports progress.
-        /// </summary>
-        public bool HasProgress =>
-            BattlesWon > 0 || BattlesLost > 0 || BattlesDrawn > 0
-            || CompletedChallenges.Count > 0 || ProficiencyGains.Count > 0 || OpenedProficiencies.Count > 0;
-
-        /// <summary>An empty summary: away time recorded but nothing simulated (a sub-threshold return).</summary>
-        public static OfflineProgressSummary Empty(long awayMs, bool autoChallengeBoss, int zoneId) => new()
-        {
-            AwayMs = awayMs,
-            AutoChallengeBoss = autoChallengeBoss,
-            ZoneId = zoneId,
-            BattlesWon = 0,
-            BattlesLost = 0,
-            BattlesDrawn = 0,
-            TotalExp = 0,
-            LevelsGained = 0,
-            StatPointsGained = 0,
-            CompletedChallenges = [],
-            ProficiencyGains = [],
-            OpenedProficiencies = [],
-        };
-    }
-
-    /// <summary>
-    /// The rewards a simulated away window applied, returned by <see cref="BattleService"/>'s offline-rewards
-    /// pass: the challenges completed and the folded proficiency gains (XP/levels/milestones/skills) plus opened
-    /// nodes. Both feed the welcome-back summary; the per-challenge and per-battle live pushes are suppressed.
-    /// </summary>
-    public record OfflineRewards(
-        IReadOnlyList<CompletedChallenge> CompletedChallenges,
-        ProficiencyAccrualResult ProficiencyGains)
-    {
-        /// <summary>No rewards: nothing was simulated in the window.</summary>
-        public static OfflineRewards Empty { get; } = new([], ProficiencyAccrualResult.Empty);
-    }
 }
