+++ docs/backend-battle.md
@@ -83,7 +83,7 @@ A skill's direct hit can carry **multiple leaf types with a split**: it owns a l
 
 ## Offline rewards simulation
 
-When a player returns after being away, `BattleService.SimulateOfflineProgress` (driven by the `GetOfflineProgress` socket command) replays the missed idle/boss battles and applies the rewards inline — the away window is bounded (≥5-minute floor, 10-hour cap) so the worst case sits well under the command timeout, and the command's cancellation token is threaded into the simulation loop. The player-facing mechanic is in [game-design.md](./game-design.md#offline-rewards); the path-specific points:
+When a player returns after being away, `OfflineProgressService.SimulateOfflineProgress` (driven by the `GetOfflineProgress` socket command) replays the missed idle/boss battles and applies the rewards inline — the away window is bounded (≥5-minute floor, 10-hour cap) so the worst case sits well under the command timeout, and the command's cancellation token is threaded into the simulation loop. `OfflineProgressService` and the live `BattleService` share zone viability/relocation and boss-zone resolution through `ZoneResolutionService` (#1516), so the two loops can never resolve a zone differently. The player-facing mechanic is in [game-design.md](./game-design.md#offline-rewards); the path-specific points:
 
 - **The simulation is parity-irrelevant.** The client never replays these battles — the backend is the sole authoritative simulator (`OfflineProgressSimulator`, pure `Game.Core`, reusing `BattleFactory`/`BattleSimulator`/`DefeatRewards`) — so there is no parity surface to mirror. Results stay consistent with online battles by reusing the shared simulator.
 - **The loop mode + zone come from the durable aggregate, not the session.** `Player.AutoChallengeBoss` (+ the current zone) selects idle vs. boss; a retired/locked/bossless boss target falls back to idle, and a non-viable idle zone is lazily relocated the same way the live idle loop relocates.
