+++ docs/backend.md
@@ -162,7 +162,7 @@ A server push that **faults** is escalated rather than silently dropped: the poi
 
 Proficiency XP is computed **server-side at battle completion, off the tick loop** — like statistics and challenge rewards, never inside the deterministic 40 ms simulation. This is the deliberate boundary that keeps it off the frontend/backend battle-parity surface: the simulation produces the `BattleStats` (the typed **output book** — `TypedDamageDealt` — and **incoming book** — `TypedDamageExposure`) and the player's power is the same `DefeatRewards` measure already computed server-side; the accrual reads those after the fact. Only the eventual proficiency *attribute bonus* is parity-sensitive (both sides compute it from player state at snapshot time), not the XP. The player-facing mechanic — the per-path power-normalized claim, two-book routing, and frontier walk — is owned by [game-design.md](./game-design.md#proficiency-xp--effect-based-power-normalized-accrual); this section is the server-side execution path only.
 
-The accrual lives in one place — `ProficiencyRewardService` (mirroring `ChallengeRewardService`) — and **both** completion paths call it, so live and offline accrue identically: the live path runs it in `BattleStatisticsEventHandler` (reacting to `BattleCompletedEvent`, whose `PlayerPower` the victory path threads on), and the offline consolidation runs it per won battle in `BattleService`. It folds the two books onto each path's `ActivityKey` (`ActivityKeys.ForDamageKey` / `ForDamageKeyResist`) and has each path claim its XP via `ProficiencyXpCalculator` against its **frontier tier** (`IProficiencies.GetPath(id).Frontier`, read off the player's current proficiency levels). The `ActivityKey → paths` routing is a reverse index the proficiency cache builds, with **retired paths excluded** so a retired track is frozen for accrual. Because the frontier is re-resolved from progress each battle, the offline batch — replaying the same loadout over the away window — evolves identically to live. It writes through the same `PlayerProgress` aggregate / write-behind path as statistics and challenges (absolute values, so a re-applied event converges). The live path raises a `ProficiencyXpGainedEvent` that an API-layer notifier pushes to the client exactly like the challenge-completion push; the offline path suppresses the push and instead **folds each won battle's accrual across the away window** (`ProficiencyGainAccumulator` — XP summed per proficiency, level/residual XP final, milestones/skills/opened nodes unioned) onto the **welcome-back summary**, which carries the gains and opened nodes (reusing the live push's models so the gate renders offline and live the same way) and counts them in `HasProgress`.
+The accrual lives in one place — `ProficiencyRewardService` (mirroring `ChallengeRewardService`) — and **both** completion paths call it, so live and offline accrue identically: the live path runs it in `BattleStatisticsEventHandler` (reacting to `BattleCompletedEvent`, whose `PlayerPower` the victory path threads on), and the offline consolidation runs it per won battle in `BattleService`. It folds the two books onto each path's `ActivityKey` (`ActivityKeys.ForDamageKey` / `ForDamageKeyResist`) and has each path claim its XP via `ProficiencyXpCalculator` against its **frontier tier** (`IProficiencies.GetPath(id).Frontier`, read off the player's current proficiency levels). The `ActivityKey → paths` routing is a reverse index the proficiency cache builds, with **retired paths excluded** so a retired track is frozen for accrual. Because the frontier is re-resolved from progress each battle, the offline batch — replaying the same loadout over the away window in `OfflineProgressService` — evolves identically to live. It writes through the same `PlayerProgress` aggregate / write-behind path as statistics and challenges (absolute values, so a re-applied event converges). The live path raises a `ProficiencyXpGainedEvent` that an API-layer notifier pushes to the client exactly like the challenge-completion push; the offline path suppresses the push and instead **folds each won battle's accrual across the away window** (`ProficiencyGainAccumulator` — XP summed per proficiency, level/residual XP final, milestones/skills/opened nodes unioned) onto the **welcome-back summary**, which carries the gains and opened nodes (reusing the live push's models so the gate renders offline and live the same way) and counts them in `HasProgress`.
 
 ### Milestone & open effects (derived openness, persisted only as skill grants)
